+++ app/src/main/java/com/xposed/doupp/MainHook.kt
@@ -45,7 +45,7 @@ class MainHook : IXposedHookLoadPackage {
     companion object {
         const val LOG_TAG = "Dou+"
         const val TARGET_PACKAGE = "com.ss.android.ugc.aweme"
-        const val MODULE_VERSION = "3.0.1"
+        const val MODULE_VERSION = "3.0.0"
     }
 
     override fun handleLoadPackage(lpparam: XC_LoadPackage.LoadPackageParam) {